% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoreyGenbank.R
\name{DoreyGenbank}
\alias{DoreyGenbank}
\title{Download larger amounts of genbank data into R}
\usage{
DoreyGenbank(
  db = "nucleotide",
  term = "(cytb[Gene Name]) AND (Apis[Organism])",
  retmax = 500,
  config = NULL,
  retmode = "xml",
  use_history = FALSE,
  seq.names = NULL,
  species.names = TRUE,
  as.character = FALSE,
  chunk.size = 400,
  quiet = TRUE,
  type = "DNA"
)
}
\arguments{
\item{db}{character, name of the database to search for.}

\item{term}{character, the search term. The syntax used in making these searches is described in the Details of this help message, the package vignette and reference given below.}

\item{config}{vector configuration options passed to httr::GET}

\item{retmode}{character, one of json (default) or xml. This will make no difference in most cases.}

\item{use_history}{logical. If TRUE return a web_history object for use in later calls to the NCBI}

\item{seq.names}{the names to give to each sequence; by default the accession numbers are used. CURRENTLY NOT IN USE}

\item{species.names}{a logical indicating whether to attribute the species names to the returned object.}

\item{as.character}{a logical controlling whether to return the sequences as an object of class "DNAbin" (the default).}

\item{chunk.size}{the number of sequences downloaded together (see details).}

\item{quiet}{a logical value indicating whether to show the progress of the downloads. If TRUE, will also print the (full) name of the FASTA file containing the downloaded sequences.}

\item{type}{a character specifying to download "DNA" (nucleotide) or "AA" (amino acid) sequences.}
}
\value{
A list of DNA sequences made of vectors of class "DNAbin", or of single characters (if as.character = TRUE) with two attributes (species and description).
}
\description{
This function works the same as \code{rentrez::entrez_search()} and \code{ape::read.GenBank()} together
to download data from Genbank. However, it overcomes issues of download limits by iteratively
downloading data and adding sequences and associated metadata together.
}
\examples{

TEST <- DoreyGenbank(db = "nucleotide",
term = "(cytb[Gene Name]) AND (Apis[Organism])",
retmax = 500)



}
